split logic into separate components
